pyproject.toml
[project]
name = "tdsprj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "markdown>=3.8",
    "numpy>=2.2.6",
    "openai>=1.83.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.3",
]
